Made KundeRenderComponent with all needed information

+++ Anden_SemesterProjekt/Client/Pages/KundePage.razor.cs
@@ -2,6 +2,7 @@
 using Anden_SemesterProjekt.Client.Services;
 using Anden_SemesterProjekt.Shared.Models;
 using Microsoft.AspNetCore.Components;
+using Microsoft.JSInterop;
 
 namespace Anden_SemesterProjekt.Client.Pages
 {
@@ -11,6 +12,9 @@ namespace Anden_SemesterProjekt.Client.Pages
 
         [Inject]
         public IKundeClientService KundeService { get; set; }
+
+        [Inject] 
+        private IJSRuntime JS { get; set; }
 
         protected override async Task OnInitializedAsync()
         {
@@ -29,8 +33,21 @@ namespace Anden_SemesterProjekt.Client.Pages
 
         private async Task OnDeleteKunde(Kunde kunde)
         {
-            await KundeService.DeleteKunde(kunde.KundeId);
-            kunder = await KundeService.GetKunder();
+            var confirmDelete = await JS.InvokeAsync<bool>("confirm", "Er du sikker på, at du vil slette denne kunde?");
+
+            if (confirmDelete)
+            {
+                var result = await KundeService.DeleteKunde(kunde.KundeId);
+
+                if (result.IsSuccessStatusCode)
+                {
+                    kunder = await KundeService.GetKunder();
+                }
+                else
+                {
+                    await JS.InvokeVoidAsync("alert", "Der skete en fejl under sletning af kunden.");
+                }
+            }
         }
     }
 }
